# Conflicts:
#	Console/obj/Debug/DesignTimeResolveAssemblyReferencesInput.cache

#	ProjetPrincipal.Data/obj/x86/Debug/DesignTimeResolveAssemblyReferencesInput.cache

#	ProjetPrincipal.Data/obj/x86/Debug/ProjetPrincipal.Data.csproj.FileListAbsolute.txt

#	ProjetPrincipal.Data/obj/x86/Debug/ProjetPrincipal.Data.csprojResolveAssemblyReference.cache
#	ShootToThrill.v12.suo

#	ShootToThrill/ShootToThrill/obj/x86/Debug/DesignTimeResolveAssemblyReferencesInput.cache

#	ShootToThrill/ShootToThrill/obj/x86/Debug/ShootToThrill.csproj.FileListAbsolute.txt

#	ShootToThrill/ShootToThrillContent/obj/x86/Debug/DesignTimeResolveAssemblyReferencesInput.cache

#	ShootToThrill/ShootToThrillContent/obj/x86/Debug/ShootToThrillContent.contentproj.FileListAbsolute.txt

#	ShootToThrill/ShootToThrillContent/obj/x86/Debug/ShootToThrillContent.contentprojResolveAssemblyReference.cache

+++ ProjetPrincipal.Data/DescriptionFusil.cs
@@ -33,7 +33,7 @@ namespace ProjetPrincipal.Data
         public float Cadence { get;  set; }
         public int Dommage { get;  set; }
         public float Portée { get;  set; }
-        public float AngleDeTir { get;  set; }
+        public int AngleDeTir { get;  set; }
         public bool Area { get;  set; }
         public int NbBallesParTir { get; set; }
     }
